requirements.txt
anyio==4.12.1
blinker==1.9.0
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
Flask==3.1.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-telegram-bot==22.2
Werkzeug==3.1.6
python-dotenv>=1.0.0
